Reference profiles from tumor-infiltrating non-malignant cells

A dataset providing reference gene expression profiles of main non-malignant cell types infiltrating solid tumours, to be used with the EPIC deconvolution method (and wrapped by IOBR). Expression values are based on single-cell RNA-seq data from melanoma tumour samples, and are given in TPM units.

Usage

data(TRef)

Format

A list with three components:

refProfiles

Numeric matrix (nGenes × nCellTypes) containing the reference gene expression (TPM counts) for each gene in each cell type. Row names are gene identifiers, column names are cell types (e.g. B cells, CAFs, CD4 T, CD8 T, endothelial cells, macrophages, NK cells).

refProfiles.var

Numeric matrix with the same dimensions as refProfiles, giving the variability (e.g. variance) of gene expression for each gene–cell-type combination. This matrix is used as weights in the EPIC optimisation.

sigGenes

Character vector listing the signature genes used by EPIC to deconvolve cell-type proportions. Only these genes are used in the fit, although refProfiles usually stores all available genes.

Source

Single-cell RNA-seq data of tumour-infiltrating non-malignant cells from melanoma (GEO accession GSE72056, 9 donors).

References

Tirosh et al. (2016) Science 352:189-196. doi:10.1126/science.aad0501 Racle et al. (2017) eLife 6:e26476. doi:10.7554/eLife.26476