Generate Reference Signature Matrix Using Limma

Performs differential expression analysis using the limma package to identify significantly expressed genes across different cell types. Computes median expression levels of these significant genes to create a reference signature matrix.

Usage

generateRef_limma(dat, pheno, FDR = 0.05)

Arguments

dat

Matrix or data frame. Gene probes in rows and samples in columns.

pheno

Character vector. Cell type class of the samples.

FDR

Numeric. Genes with BH adjusted p-value < FDR are considered significant. Default is 0.05.

Value

List containing: - `reference_matrix`: Data frame of median expression values for significantly expressed genes. - `G`: Number of probes used that resulted in the lowest condition number. - `condition_number`: Minimum condition number obtained. - `whole_matrix`: Matrix of median values across all samples.

Examples

dat <- matrix(rnorm(2000), nrow = 100)
rownames(dat) <- paste("Gene", 1:100, sep = "_")
colnames(dat) <- paste("Sample", 1:20, sep = "_")
pheno <- sample(c("Type1", "Type2", "Type3"), 20, replace = TRUE)
# \donttest{
results <- generateRef_limma(dat, pheno)
print(results)
#> $reference_matrix
#> [1] Type1 Type2 Type3
#> <0 rows> (or 0-length row.names)
#> 
#> $G
#> [1] 50
#> 
#> $condition_number
#> [1] 0
#> 
#> $whole_matrix
#>                 Type1        Type2        Type3
#> Gene_1   -0.294228071 -0.015381692 -0.183292667
#> Gene_2    0.400736372  0.171354918  0.823668020
#> Gene_3    0.552943994 -0.601546123 -0.292481786
#> Gene_4    0.132534263  0.872542832 -1.187090448
#> Gene_5    0.582163199  0.140168505  0.976808879
#> Gene_6    0.407546075 -0.157363602  0.240330467
#> Gene_7   -0.244828597  0.430500319  0.918647834
#> Gene_8    0.481699743  0.287452218 -0.465588037
#> Gene_9    0.078587185 -0.442087478  0.104257989
#> Gene_10   0.368188221  0.566943146  0.327411498
#> Gene_11  -0.567112162 -0.746501813  0.152063451
#> Gene_12   0.005214646 -0.651919026 -0.381171789
#> Gene_13  -0.074808505  0.424446591 -0.537401224
#> Gene_14   0.093065117 -0.203097514  0.191643407
#> Gene_15   0.334862783 -0.391630356  0.564117698
#> Gene_16  -0.098715350 -0.157465989 -0.383801745
#> Gene_17  -0.722761109 -0.429447297 -0.222144279
#> Gene_18   0.257764633  0.389543077 -0.815077103
#> Gene_19   0.016526801  0.294285597  0.499524038
#> Gene_20  -0.923309912  0.137946418 -0.093489069
#> Gene_21   0.427589508  0.831400019 -0.558431176
#> Gene_22  -0.135720718 -0.337900122 -0.202897698
#> Gene_23  -0.010601259 -1.236030816 -0.168853280
#> Gene_24   0.169489632 -0.199489570  0.230113871
#> Gene_25  -0.657383490  0.116640434 -0.623073907
#> Gene_26   0.650189192  0.746920846  0.185232995
#> Gene_27  -0.039526397  0.131083420 -0.462425013
#> Gene_28  -0.742315724 -0.216706527  0.446609954
#> Gene_29  -0.302499456  0.610810758 -0.457662959
#> Gene_30   0.155214248 -0.066387947  0.632310731
#> Gene_31  -0.585260175 -0.818668986 -0.093587464
#> Gene_32   0.126316908  0.091095477 -0.273687202
#> Gene_33  -0.412876368 -0.226381775 -0.724520329
#> Gene_34  -0.507510884  0.173141537 -0.924954561
#> Gene_35  -0.023550649 -0.232096799  0.552963211
#> Gene_36  -0.754596612 -0.146001362 -0.172250354
#> Gene_37   0.104656476  0.009026432  0.505257106
#> Gene_38   0.286030001 -0.728951475  0.009068848
#> Gene_39   0.546357655  0.290548346 -0.432816779
#> Gene_40   0.542126060  0.749164704  0.641851483
#> Gene_41   0.303301533  0.240569663 -0.047112463
#> Gene_42  -0.667863520  0.151596351 -0.076736280
#> Gene_43   0.025763003  0.244023674 -0.282018128
#> Gene_44  -0.676889138  0.060043704 -0.047156556
#> Gene_45  -0.095658713  0.233804483 -0.690401864
#> Gene_46   0.984974173 -0.587512791 -0.546358727
#> Gene_47  -0.497475732  0.394251963 -0.318685552
#> Gene_48  -0.583531336 -0.242784224 -0.209704374
#> Gene_49   0.118350974  0.490081357  0.127516355
#> Gene_50   0.296873942  0.021298983  0.589201053
#> Gene_51   0.093262020  0.367893530  0.089164914
#> Gene_52  -0.250679496  0.603555522 -0.299958876
#> Gene_53   0.826404223 -0.487864028  0.372147359
#> Gene_54  -0.230859994 -0.077540473  0.411797309
#> Gene_55   0.040972595  0.054492976  0.056567043
#> Gene_56  -0.065177513 -0.109977143  0.165934124
#> Gene_57  -0.258041038 -0.701906079 -0.410226024
#> Gene_58  -0.298349351 -0.166392473 -0.536677767
#> Gene_59  -0.254416939  0.292595813  0.363648591
#> Gene_60  -0.949641782 -0.549346194  0.207599409
#> Gene_61  -0.312827414 -0.276206744 -0.178981854
#> Gene_62  -0.163945983 -0.240739682 -1.234878892
#> Gene_63  -0.272837303 -0.055322094 -0.082495409
#> Gene_64   0.051520009 -0.346723172 -0.125363892
#> Gene_65  -0.300849294  0.178318644  0.305354059
#> Gene_66  -0.126956208 -1.274123893 -0.214067093
#> Gene_67   0.044676079 -0.099623305 -0.372650997
#> Gene_68  -0.216702235  0.443627660 -0.517181092
#> Gene_69  -0.074152424 -0.247850748  0.073699177
#> Gene_70   0.191185802 -0.435639370  0.109643844
#> Gene_71   0.562186089  0.239675658  0.036785870
#> Gene_72   0.314591956  0.474168441 -0.460950033
#> Gene_73   0.095804898 -0.877017463  0.697954494
#> Gene_74  -0.452727671  0.146558730 -0.332863128
#> Gene_75  -0.667533062 -0.447295957 -0.510226788
#> Gene_76   0.176410479 -0.381839585  0.298287826
#> Gene_77   0.562434281  0.575552359 -0.473239379
#> Gene_78  -0.024229187 -0.640793822 -0.235575229
#> Gene_79  -0.077093785  0.487962211  0.222612698
#> Gene_80   0.127942084  0.925939689 -0.036237954
#> Gene_81   0.529814625 -0.139746067 -0.644962454
#> Gene_82   0.008995257  0.443023984 -0.356074985
#> Gene_83   0.011096222  0.410211477 -0.284243620
#> Gene_84   0.529205574 -0.111895973  0.590881840
#> Gene_85   0.604381516 -0.222546394  0.153118697
#> Gene_86  -0.373755700  0.584248357 -0.125484249
#> Gene_87  -0.431805402 -0.815425841 -0.525794933
#> Gene_88   0.459090364 -0.623467492  0.368832393
#> Gene_89  -0.010414339  0.063652895 -0.612202213
#> Gene_90  -0.206138394  0.888882708  0.039132378
#> Gene_91  -0.356563349 -0.151753297 -0.655177863
#> Gene_92   0.335824923 -0.059523836 -0.342026590
#> Gene_93  -1.020944906  0.277786950 -0.893571204
#> Gene_94  -0.239075948  0.370658131 -0.356783849
#> Gene_95  -0.440602341  0.399104916  0.186282442
#> Gene_96   0.335746589  0.453475347  0.014851549
#> Gene_97   0.205839352 -0.342692041  0.473058842
#> Gene_98  -0.467696117  0.741255244  0.607729635
#> Gene_99   0.255420694  0.171132601  0.095113744
#> Gene_100 -0.607378439  0.336751194  0.538853930
#> 
# }