Reference profiles for immune cell types using lm22 (EPIC/IOBR)

lm22 is a gene expression signature matrix used in the EPIC and IOBR packages for estimating the proportions of different immune cell types from bulk RNA-seq data. It contains reference profiles for multiple immune cell types, including various B cell and T cell subsets, as well as plasma cells and other immune cell types.

Usage

data(lm22)

Format

A list of length 2:

refProfiles

A numeric matrix where rows represent genes and columns represent different immune cell types. The matrix provides gene expression profiles used as references for deconvolution in the EPIC model. The number of columns may vary depending on the number of immune cell types included in the reference profiles.

cellTypes

A character vector containing the names of the immune cell types included in the reference profiles.

Examples

data(lm22)
head(lm22)
#>        B cells naive B cells memory Plasma cells T cells CD8 T cells CD4 naive
#> ABCB4      555.71345       10.74423     7.225819     4.31128           4.60586
#> ABCB9       15.60354       22.09479   653.392328    24.22372          35.67151
#> ACAP1      215.30595      321.62102    38.616872  1055.61338        1790.09717
#> ACHE        15.11795       16.64885    22.123737    13.42829          27.18773
#> ACP5       605.89738     1935.20148  1120.104684   306.31252         744.65660
#> ADAM28    1943.74270     1148.12014   324.780800    22.68972          40.06171
#>        T cells CD4 memory resting T cells CD4 memory activated
#> ABCB4                    7.406442                     8.043976
#> ABCB9                   30.048192                    38.455423
#> ACAP1                  922.152747                   340.883424
#> ACHE                    18.444927                    13.441268
#> ACP5                   557.819820                   248.546932
#> ADAM28                  21.233311                    15.209620
#>        T cells follicular helper T cells regulatory (Tregs) T cells gamma delta
#> ABCB4                   6.469993                   7.833082            9.312295
#> ABCB9                  17.604795                  46.073657           19.715722
#> ACAP1                1107.798265                1995.482601          280.075719
#> ACHE                   14.805542                  24.652712           33.658448
#> ACP5                  711.949744                 958.916018          493.969120
#> ADAM28                 24.281406                  74.766335           24.370752
#>        NK cells resting NK cells activated  Monocytes Macrophages M0
#> ABCB4          13.68284           4.556713   6.202496        7.95136
#> ABCB9          21.86372          16.396143  20.309426       56.25189
#> ACAP1         512.63691         387.581742 190.879024       48.97222
#> ACHE           16.65374          13.441268  16.375268       13.69296
#> ACP5          340.82014         239.277068 512.931382    13644.17230
#> ADAM28        193.60939          94.405818  58.311203       54.96649
#>        Macrophages M1 Macrophages M2 Dendritic cells resting
#> ABCB4        27.55771      121.43228               114.63675
#> ABCB9        29.38088       40.39026               123.99331
#> ACAP1        30.97425       23.05515                61.38655
#> ACHE       1373.06962       48.58576                16.71547
#> ACP5       1062.47591     7633.96018             13825.05363
#> ADAM28      274.64000      360.68204                37.29173
#>        Dendritic cells activated Mast cells resting Mast cells activated
#> ABCB4                   4.872184           4.727985             34.11366
#> ABCB9                  23.384603          53.466757             23.61575
#> ACAP1                  26.412923         104.200198             73.60793
#> ACHE                  193.159977          13.948778             20.50125
#> ACP5                 3989.865169         263.497930            323.38128
#> ADAM28                 53.219649         137.853628            116.87950
#>        Eosinophils Neutrophils
#> ABCB4     15.03053    14.90689
#> ABCB9     29.78644    33.67915
#> ACAP1    696.44280   596.02596
#> ACHE      40.41493    22.76649
#> ACP5     860.56337   307.14280
#> ADAM28   546.00616   128.68607